Add simple profiling, change geometry indices to uint32 from ushort

+++ src/core/components/collision/collider.h
@@ -128,7 +128,7 @@ public:
         
         int attrCount = meshData->GetAttribCount<Au::Position>();
         Au::Position* posData = meshData->GetAttribData<Au::Position>();
-        std::vector<unsigned short>& indices = meshData->GetIndices();
+        std::vector<unsigned>& indices = meshData->GetIndices();
         if(!posData || !attrCount)
             return;
